fix reset key ignored bug
For reasons that are not understood, when the -O3 optimized version of
the app are run in auto-start mode the reset key push is not detected
most of the time unless a delay is added. Oddly, this bug doesn't
appear if the optimized version is run from the command line, which
makes no sense. It doesn't appear to be code getting optimized away or
a missing volatile declaration but rather timing related.